Added "Remember me" functionality and fixed login page.

+++ source/account/views.py
@@ -5,7 +5,6 @@ from django.contrib.auth.models import User
 from django.contrib.auth import authenticate, login, logout
 from django.views.generic.base import View
 from django.views.generic.edit import FormView
-# from django.contrib.auth.decorators import login_required
 
 from .forms import UserForm
 from django import forms
@@ -18,13 +17,20 @@ def login_view(request):
     user_form = UserForm()
 
     if request.method == 'POST':
-        if request.session.test_cookie_worked():
-            # if user_form.cleaned_data[]
-            request.session.delete_test_cookie()
-
         user_form = UserForm(request.POST)
 
         if user_form.is_valid():
+            if 'remember_me' in user_form.cleaned_data:
+                if request.session.test_cookie_worked():
+                    request.session.delete_test_cookie()
+                
+                    if user_form.cleaned_data['remember_me']:
+                        request.session.set_expiry(settings.REMEMBER_ME_EXPIRY)
+                    else:
+                        request.session.set_expiry(0)
+                    
+                    user_form.cleaned_data.pop('remember_me')
+
             user = authenticate(**user_form.cleaned_data)
 
             if user is not None:
@@ -32,8 +38,9 @@ def login_view(request):
                 return redirect('account:index')
 
     if settings.REMEMBER_ME:
-        request.session.set_expiry(settings.REMEMBER_ME_EXPIRY)
         request.session.set_test_cookie()
+
+    print(request.get_host())
 
     return render(request, 'account/login.html', {'user_form': user_form})
